+++ packages/fury-adapter-oas3-parser/test/unit/parser/oas/parseOperationObject-test.js
@@ -293,6 +293,7 @@ describe('Operation Object', () => {
             {
               name: 'resource',
               in: 'path',
+              required: true,
             },
           ],
           responses: {},