+++ packages/fury-adapter-oas3-parser/test/unit/parser/oas/parseParameterObjects-test.js
@@ -24,10 +24,12 @@ describe('Parameter Objects', () => {
       {
         name: 'id',
         in: 'path',
+        required: true,
       },
       {
         name: 'slug',
         in: 'path',
+        required: true,
       },
       {
         name: 'tags',